Added changes in course update and deletion via auth

+++ controllers/courseController.js
@@ -60,17 +60,23 @@ const updateCourse = async (req, res) => {
       return res.status(404).json({ message: 'Course not found' });
     }
 
-    // Check if audiobooks exist before updating the course
-    if(req.body.contents)
-    {  
-      let audiobooksExist = await Audiobook.find({ _id: { $in: req.body.contents } });
-      if (audiobooksExist.length !== contents.length) {
-        return res.status(400).json({ message: 'One or more audiobooks do not exist' });
-      }
+    if(course.createdBy != req.user.id)
+    {
+      res.status(401).json({message : "This Course doesn't belong to you. You can edit your own Courses"});
     }
-    const updateCourse = await Course.findByIdAndUpdate(req.params.id,req.body);
+    else{
+        if(req.body.contents)
+      {      
+        // Check if audiobooks exist before updating the course
+        let audiobooksExist = await Audiobook.find({ _id: { $in: req.body.contents } });
+        if (audiobooksExist.length !== contents.length) {
+          return res.status(400).json({ message: 'One or more audiobooks do not exist' });
+        }
+      }
+      const updateCourse = await Course.findByIdAndUpdate(req.params.id,req.body);
 
-    res.json({message : "Course Updated",data: updateCourse});
+      res.json({message : "Course Updated",data: updateCourse});
+    }
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
@@ -83,8 +89,15 @@ const deleteCourse = async (req, res) => {
       return res.status(404).json({ message: 'Course not found' });
     }
 
-    await Course.deleteOne({_id:req.params.id});
-    res.json({ message: 'Course deleted successfully' });
+    if(course.createdBy != req.user.id)
+    {
+      res.status(401).json({message : "This Course doesn't belong to you. You can edit your own Courses"});
+    }
+    else
+    {
+      await Course.deleteOne({_id:req.params.id});
+      res.json({ message: 'Course deleted successfully' });
+    }
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
